refactor: Rename `lib/test/objectSubsetEquals.js`
