+++ lib/test/object-subset-equals.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const objectSubsetEquals = require('./objectSubsetEquals');
+const objectSubsetEquals = require('./object-subset-equals');
 
 describe('objectSubsetEquals', () => {
   it('correctly compares equal primitives', () => {
