+++ lib/test/runTest.js
@@ -3,7 +3,7 @@
 const { entries } = require('lodash');
 const fetch = require('node-fetch');
 const { TestError } = require('./errors');
-const objectSubsetEquals = require('./objectSubsetEquals');
+const objectSubsetEquals = require('./object-subset-equals');
 
 const runTest = async (testSpec, path, method, baseApiUrl) => {
   let body;
